Create Unity Catalog
In this project we will familiarise ourselves with the Unity Catalog tool by conducting a hands-on lab. We will review its basic concepts and develop the necessary steps to install, configure and use this tool that helps us to govern and collaborate around data and data products, such as artificial intelligence models.